requirements.txt
asgiref==3.5.2
Django==4.1.2
pip==21.1.2
PyMySQL==1.0.2
PyYAML==6.0
setuptools==57.0.0
sqlparse==0.4.3
tzdata==2022.5
wheel==0.36.2
pylint==2.12.2
pylint-django==2.5.3
pyjwt==2.6.0
minio==7.0.4
pandas==1.5.1
pyecharts==1.9.1
cryptography
